+++ anan-cloudadviced/anan-platformserver/src/main/java/top/fosin/anan/platform/modules/pay/query/PayOrderQuery.java
@@ -51,7 +51,7 @@ public class PayOrderQuery extends LogiSortQuery<LogiQueryRule, SortRule, Long> 
     private Date orderTime;
 
     @ApiModelProperty(value = "订单状态：0=新建，1=支付，2=取消，3=作废", example = "Integer")
-    private Integer status;
+    private Byte status;
 
     @DateTimeFormat(pattern = DateTimeUtil.DATETIME_PATTERN)
     @ApiModelProperty(value = "支付日期", example = "Date")
