+++ anan-cloudadviced/anan-platformserver/src/main/java/top/fosin/anan/platform/modules/role/dto/RoleDTO.java
@@ -37,7 +37,7 @@ public class RoleDTO extends Id<Long> {
     private String tips;
 
     @ApiModelProperty(value = "使用状态：0=启用，1=禁用，具体取值于字典表anan_dictionary.code=11", required = true, example = "Integer")
-    private Integer status;
+    private Byte status;
 
     @DateTimeFormat(pattern = DateTimeUtil.DATETIME_PATTERN)
     @ApiModelProperty(value = "创建时间", required = true, example = "Date")
@@ -54,6 +54,6 @@ public class RoleDTO extends Id<Long> {
     private Long updateBy;
 
     @ApiModelProperty(value = "内置标志：是否是系统内置角色，内置角色不能被用户删除和修改，0=不是 1=是", required = true, example = "Integer")
-    private Integer builtIn;
+    private Byte builtIn;
 
 }
